PER/requirements.txt
# ==============================================================================
# Requirements pour PER - Benchmark ML vs Transformers
# Classification de logs CI/CD
# ==============================================================================

# Core
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0

# Deep Learning
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.21.0

# Évaluation et métriques
evaluate>=0.4.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilitaires
tqdm>=4.65.0
psutil>=5.9.0
memory-profiler>=0.61.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# Formatage et qualité
black>=23.0.0
flake8>=6.0.0

# Tests
pytest>=7.0.0
